+++ mrb_doc/core/window.rb
@@ -1,12 +1,3 @@
-# Sets the monitor for the current window.
-# @param monitor [Integer]
-# @return [nil]
-def set_window_monitor(monitor)
-  # mrb_set_window_monitor
-  # src/mruby_integration/core/window.cpp
-  nil
-end
-
 # Gets the scale of the window.
 # @return [Vector2]
 def get_window_scale_dpi